MB5.2 Hernoemen bestanden en aanpassen van Integer naar long en int

+++ src/main/java/be/vdab/scrumjava202406/bestellingen/ArtikelNietGevondenException.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.NOT_FOUND)
 class ArtikelNietGevondenException extends RuntimeException {
-    ArtikelNietGevondenException(Integer id) {
+    ArtikelNietGevondenException(long id) {
         super("Artikel niet gevonden. Id: " + id);
     }
 }
